Create abstract superclass Contact, from which Phone, Email and Address extend
