+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,42 @@
+package seedu.addressbook.data.person;
+
+import seedu.addressbook.data.exception.IllegalValueException;
+
+/**
+ * Represents a Person's contact in the address book.
+ */
+public abstract class Contact {
+    public String value;
+    protected boolean isPrivate;
+
+    public Contact() {
+        this("", false);
+    }
+
+    /**
+     * Validates given address.
+     *
+     * @throws IllegalValueException if given address string is invalid.
+     */
+    public Contact(String address, boolean isPrivate) {
+        this.value = address;
+        this.isPrivate = isPrivate;
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public abstract boolean equals(Object other);
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
